# wazevo: fixes accidental heap allocation in VarLengthPool.Allocate