```
>> benchstat old.txt new8.txt
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/stdlibs
                             │   old.txt   │              new8.txt               │
                             │   sec/op    │   sec/op     vs base                │
Zig/Compile/test-opt.wasm-10   4.750 ± ∞ ¹   4.842 ± ∞ ¹       ~ (p=1.000 n=1) ²
Zig/Compile/test.wasm-10       5.931 ± ∞ ¹   5.899 ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                        5.308         5.345        +0.69%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                             │    old.txt    │               new8.txt                │
                             │     B/op      │     B/op       vs base                │
Zig/Compile/test-opt.wasm-10   464.0Mi ± ∞ ¹   455.0Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
Zig/Compile/test.wasm-10       717.8Mi ± ∞ ¹   703.8Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                        577.1Mi         565.9Mi        -1.94%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                             │    old.txt    │               new8.txt                │
                             │   allocs/op   │  allocs/op    vs base                 │
Zig/Compile/test-opt.wasm-10    953.9k ± ∞ ¹   436.8k ± ∞ ¹        ~ (p=1.000 n=1) ²
Zig/Compile/test.wasm-10       1137.6k ± ∞ ¹   555.1k ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                         1.042M         492.4k        -52.73%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```